+++ recipes/configure-capture-agent-manager-nginx-proxy.rb
@@ -4,8 +4,8 @@
 include_recipe "mh-opsworks-recipes::update-package-repo"
 ::Chef::Recipe.send(:include, MhOpsworksRecipes::RecipeHelpers)
 
-capture_agent_manager_info = node.fetch(:capture_agent_manager, {})
-app_name = capture_agent_manager_info.fetch(:capture_agent_manager_name, "capture_agent_manager")
+app_name = get_capture_agent_manager_app_name
+username = get_capture_agent_manager_usr_name
 
 install_package("nginx")
 
@@ -17,20 +17,21 @@ if cert_defined(ssl_info)
   certificate_exists = true
 end
 
-directory '/etc/nginx/proxy-includes' do
-  owner 'root'
-  group 'root'
+directory "/etc/nginx/proxy-includes" do
+  owner "root"
+  group "root"
 end
 
-template %Q|/etc/nginx/sites-enabled/default| do
+template "/etc/nginx/sites-enabled/default" do
   source "nginx-proxy-ssl-only.erb"
   manage_symlink_source true
 end
 
-template %Q|/etc/nginx/proxy-includes/capture-agent-manager.conf| do
+template "/etc/nginx/proxy-includes/capture-agent-manager.conf" do
   source "nginx-proxy-capture-agent-manager.conf.erb"
   variables({
-    capture_agent_manager: app_name
+    capture_agent_manager: app_name,
+    capture_agent_manager_username: username
   })
 end