+++ recipes/create-capture-agent-manager-directories.rb
@@ -1,14 +1,18 @@
 # Cookbook Name:: mh-opsworks-recipes
 # Recipe:: create-python-manager-directories
 
+::Chef::Recipe.send(:include, MhOpsworksRecipes::RecipeHelpers)
+
+username = get_capture_agent_manager_usr_name
+
 [
-  "/home/capture_agent_manager/sites",
-  "/home/capture_agent_manager/sock",
-  "/home/capture_agent_manager/logs"
+  %Q|home/#{username}/sites|,
+  %Q|home/#{username}/sock|,
+  %Q|home/#{username}/logs|
 ].each do |capture_agent_manager_directory|
   directory capture_agent_manager_directory do
-    owner "capture_agent_manager"
-    group "capture_agent_manager"
+    owner username
+    group username
     mode "755"
     recursive true
   end