+++ recipes/create-capture-agent-manager-user.rb
@@ -1,22 +1,26 @@
 # Cookbook Name:: mh-opsworks-recipes
 # Recipe:: create-capture-agent-manager-user
 
-group "capture_agent_manager" do
+::Chef::Recipe.send(:include, MhOpsworksRecipes::RecipeHelpers)
+
+username = get_capture_agent_manager_usr_name
+
+group username do
   append true
   gid 2133
 end
 
-user "capture_agent_manager" do
+user username do
   supports manage_home: true
   comment "capture agent manager user"
   uid 2133
-  gid "capture_agent_manager"
+  gid 2133
   shell "/bin/false"
-  home "/home/capture_agent_manager"
+  home "/home/#{username}"
 end
 
-directory "/home/capture_agent_manager/.ssh" do
-  owner "capture_agent_manager"
-  group "capture_agent_manager"
+directory "/home/#{username}/.ssh" do
+  owner username
+  group username
   mode "700"
 end